Add files via upload
These files are some of the small projects I, along with friends and the help of the internet/AI, completed for my Introduction to Python class for Uni in 2024. These were one of my first Python projects. They are both interactive games. One is a memory game and the other solves Sudoku puzzles. Both of theses codes are in Portuguese, since my classes are in Portuguese.